backend/requirements.txt
Flask==3.1.0
boto3==1.35.86
requests==2.32.3
python-dotenv==1.0.1
flask_cors==3.0.10